Rework JUnit support + other minor changes
wvtool now converts "Testing" lines to JUnit <testsuite> and check lines
to JUnit <testcase>. This is probably more in line with the original JUnit
intention. Later, we may want to make the behaviour switchable. The new
output is more clear when displayed in Jenkins classic UI, but Blue Ocean
display is not that nice.

Introduce --junit-prefix command line argument.